functions/webhook-processor/requirements.txt
functions-framework==3.4.0
google-cloud-pubsub==2.20.0
flask==2.3.3